removed some superfluous files and moved the checkout form
