+++ lwb-music/src/App.js
@@ -10,6 +10,7 @@ import Music from './pages/Music';
 import About from './pages/About';
 import LandingPage from './pages/LandingPage';
 import Greeting from './components/Greeting';
+import Checkout from './pages/CheckoutPage';
 
 
 function App() {
@@ -18,6 +19,7 @@ function App() {
     <Router>
       <Layout>
       <Switch>
+        <Route exact path='/checkout' component={Checkout} />
         <Route exact path='/thankyou' component={Greeting} />
         <Route exact path='/contact' component={Contact} />
         <Route exact path='/resources' component={Resources} />
