+++ lwb-music/src/index.js
@@ -1,18 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css';
 import App from './App';
-import { ChakraProvider, extendTheme } from '@chakra-ui/react';
-
-const colors = {
-  brand: {
-    // 900: '#1a365d',
-    // 800: '#153e75',
-    // 700: '#2a69ac',
-  },
-}
-
-const theme = extendTheme({ colors })
+import { ChakraProvider } from '@chakra-ui/react';
 
 ReactDOM.render(
   <React.StrictMode>
